Combine LCA libraries
